Fixed return type of xpt2046_platform_get_millis() for nrfconnect sdk platform

+++ platforms/nrfconnectsdk_zephyr_nrf52/platform_ncs_zephyr_nrf52.c
@@ -63,7 +63,7 @@ void xpt2046_platform_delay(uint64_t ms)
     k_sleep(K_MSEC(ms));
 }
 
-uint32_t xpt2046_platform_get_millis(void)
+uint64_t xpt2046_platform_get_millis(void)
 {
     return k_uptime_get();
 }
